Break benchmarks into separate file
