+++ src/dbnode/persist/fs/msgpack/encoder_decoder_bench_test.go
@@ -1,0 +1,84 @@
+// Copyright (c) 2018 Uber Technologies, Inc
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+// THE SOFTWARE
+
+package msgpack
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/require"
+)
+
+func BenchmarkLogEntryDecoder(b *testing.B) {
+	// Copy so we don't mutate global state
+	logEntry := testLogEntry
+	logEntry.Metadata = nil
+	logEntry.Annotation = nil
+	var (
+		enc    = NewEncoder()
+		dec    = NewDecoder(nil)
+		stream = NewDecoderStream(nil)
+		err    error
+	)
+
+	require.NoError(b, enc.EncodeLogEntry(logEntry))
+	buf := enc.Bytes()
+	for n := 0; n < b.N; n++ {
+		stream.Reset(buf)
+		dec.Reset(stream)
+		_, err = dec.DecodeLogEntry()
+		if err != nil {
+			panic(err)
+		}
+	}
+}
+
+var benchmarkBuf []byte
+
+func BenchmarkLogEntryEncoderFast(b *testing.B) {
+	var (
+		logEntry = testLogEntry
+		err      error
+	)
+	benchmarkBuf = []byte{}
+
+	for n := 0; n < b.N; n++ {
+		benchmarkBuf, err = EncodeLogEntryFast(benchmarkBuf[:0], logEntry)
+		if err != nil {
+			panic(err)
+		}
+	}
+}
+
+func BenchmarkLogEntryEncoder(b *testing.B) {
+	logEntry := testLogEntry
+	var (
+		enc = NewEncoder()
+		err error
+	)
+
+	for n := 0; n < b.N; n++ {
+		enc.EncodeLogEntry(logEntry)
+		if err != nil {
+			panic(err)
+		}
+		benchmarkBuf = enc.Bytes()
+	}
+}
